Update sv.json
There's a few rows where I can't quite tell what the particular code is for, which could affect how it should be written. 
I did a few corrections to the previous Swedish text, as it was flat out wrong in some cases. 
I don't understand what attribute this is for "MAGIC.POWER": "Kraft" - in the english version it says "POWER" and I don't understand if it is related to Magic or Items of Power. "Kraft" could work but it could be better translated if I knew what its for. 

But in general, the translation is finished and I think it should work regardless.